Modernize dependencies and upgrade to Spring Boot 4.1 (#56)
* fix: preserve request IDs across responses
* build: complete Spring Boot 3.5 migration
* build: upgrade to Spring Boot 4.1
* ci: modernize dependency maintenance
* fix: keep Flyway in control of production schema
* test: keep Flyway in control of test schema

+++ src/main/java/org/jeduardo/entries/controller/EntryController.java
@@ -24,7 +24,7 @@ public class EntryController {
 
     @RequestMapping(value = "/api/v1/entries/{id}", method = RequestMethod.GET, produces = "application/json")
     @ResponseBody
-    public Entry list(@PathVariable("id") int id) {
+    public Entry list(@PathVariable("id") long id) {
         Entry entry = entryRepository.findById(id).orElse(null);
         if (entry != null) {
             LOGGER.info(String.format("Entry found: %s", entry));
@@ -46,7 +46,7 @@ public class EntryController {
 
     @RequestMapping(value = "/api/v1/entries/{id}", method = RequestMethod.POST, consumes = "application/json", produces = "application/json")
     @ResponseBody
-    public Entry update(@PathVariable("id") int id, @RequestBody Entry entry) {
+    public Entry update(@PathVariable("id") long id, @RequestBody Entry entry) {
         Entry targetEntry = entryRepository.findById(id).orElse(null);
         if (targetEntry != null) {
             LOGGER.info(String.format("Entry found: %s", entry));
@@ -61,7 +61,7 @@ public class EntryController {
 
     @RequestMapping(value = "/api/v1/entries/{id}", method = RequestMethod.DELETE, produces = "application/json")
     @ResponseBody
-    public Entry delete(@PathVariable("id") int id) {
+    public Entry delete(@PathVariable("id") long id) {
         Entry targetEntry = entryRepository.findById(id).orElse(null);
         if (targetEntry != null) {
             entryRepository.delete(targetEntry);
